Implement a basic error logging utility class.
The ErrorLogger class provides the capability of logging an error
message to any std::ostream. The ErrorLogger also keeps track of whether an
error has been encountered. Since the ErrorLogger is implemented as a
singleton, any interpreter component can use the ErrorLogger instance to
query whether an error has occurred, clear the error flag, or log an
error message.