Move Tokens away from using std::string for literals, use std::any instead.
Previously, scanned tokens stored literal values as std::string. This
decision will have a negative impact later on when we go to evaluate
expressions but find that we are missing type info. It's better to use
std::any in this case to preserve literal type info and easily cast the
std::any object to the right type when needed.

Debug info is now turned on by default. This is useful for debugging lox
in GDB. A future revision will remove the "hardcoded" debug info option
and update the CMakeLists.txt to support Release/Debug builds proper.