Simplify the Expr class hierarchy by using std::any over ValueGetter.
The Expr AST implementation uses the Visitor pattern to handle processing
of the different AST node types. We want the Visitor pattern's Accept() method
to return a value whose type can vary. Prior to this commit, the ValueGetter
class was implemented to store value data at the node. This changeset removes
the need for ValueGetter, instead, Accept() directly returns a std::any object.
The result is a reduction in code and complexity later when we access the
node value in the interpreter.