Force explicit type declarations in ast.json.
Previously, the generate_ast.py script had logic built-in where if a
subtype had a field with a type that matched the AST base name, then
that field type would be wrapped up in a std::shared_ptr. This change
removes that logic and requires that the User explicitly define the type
as std::shared_ptr<BASE_NAME> in ast.json. This is a better approach
than having undocumented rules bakes into the code.