Add an explicit SetEnclosingEnv() method to the Environment class.
Previously, the Environment class had a constructor that accepted a
pointer to the Environment's parent or enclosing Environment. This
change replaces that constructor with a setter method,
SetEnclosingEnv(), which sets the parent/enclosing Environment of this
Environment. The setter comes more in handy when setting the enclosing
environment on a pre-existing Environment object and is needed when
managing function call environments.