Apply a bugfix to Peek(), PeekNext(), and IsAtEnd().
This changeset updates Peek() and PeekNext() to do bounds checking prior
to performing the peek operation.

Previously, IsAtEnd() would check for the end of the source string by looking
for the NULL terminator under the current cursor location. The latter would
cause SEGFAULTs to occur when the cursor had been advanced beyond the end of
the source string. We change IsAtEnd()'s behavior such that it compares the
current cursor location against the source string length.