Add debug info support to the Compiler an build_lox.sh build script.
This changeset adds debug info support to the Lox build script. A new
option, -g, is introduced that when present on the command line will
build the Compiler library with debug support. When built with debug
enabled, the Compiler will print the disassembled form of each valid
chunk of code it compiles.