Add an Object type used to represent the various Lox objects.
This changeset introduces the lox::obj::Object. Object's are used to
represent different objects as encountered by the interpreter (e.g.,
functions, instances, strings, etc.). The introduction of
lox::obj::Object necessitated minor changes to Value in order to support
proper conversion of Objects to Values and vice versa.

Included in this changset is a modification to how lox::value::Value
objects store their type specific data. Instead of using an old school C
union, we now capture Value data using a std::variant. std::variant is a
good choice here because it allows the function of a union with added
type safety.