Add a custom Stack class for use in the Virtual Machine.
The VirtualMachine class previously used a std::stack data structure to
manage runtime Value objects. std::stack lacked a lot of the
functionality required by the VM of the stack. Namely, it was
inefficient to randomly index the contents of the stack. This changeset
adds a custom Stack class which supports the usual stack operations as
well as the random indexing capability required by the Virtual Machine.